requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.1.6
django-environ==0.12.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
factory_boy==3.3.3
Faker==36.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jiter==0.8.2
kombu==5.4.2
openai==1.63.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
pytest-asyncio
django-pydantic-field
